requirements.txt
absl-py==0.11.0
astor==0.8.1
certifi==2020.6.20
gast==0.4.0
grpcio==1.33.2
importlib-metadata==2.0.0
Markdown==3.3.3
numpy==1.14.5
opencv-python==4.4.0.46
protobuf==3.13.0
PyYAML==5.3.1
scipy==1.5.3
six==1.15.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
tqdm==4.51.0
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.4.0
